+++ layers/Engine/packages/component-model/src/TypeResolvers/ScalarType/AnyScalarScalarTypeResolver.php
@@ -1,0 +1,33 @@
+<?php
+
+declare(strict_types=1);
+
+namespace PoP\ComponentModel\TypeResolvers\ScalarType;
+
+use PoP\ComponentModel\Feedback\ObjectTypeFieldResolutionFeedbackStore;
+use PoP\GraphQLParser\Spec\Parser\Ast\AstInterface;
+use stdClass;
+
+class AnyScalarScalarTypeResolver extends AbstractScalarTypeResolver
+{
+    public function getTypeName(): string
+    {
+        return 'AnyScalar';
+    }
+
+    public function getTypeDescription(): ?string
+    {
+        return $this->__('Wildcard type representing any of GraphQL\'s scalar types, including built-in types (String, Int, Boolean, Float or ID) and custom types', 'component-model');
+    }
+
+    /**
+     * Accept anything and everything
+     */
+    public function coerceValue(
+        string|int|float|bool|stdClass $inputValue,
+        AstInterface $astNode,
+        ObjectTypeFieldResolutionFeedbackStore $objectTypeFieldResolutionFeedbackStore,
+    ): string|int|float|bool|object|null {
+        return $inputValue;
+    }
+}